@@ -4,7 +4,7 @@ version: 1.2.3+4
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0-259.9.beta <3.0.0'
 
 dependencies:
   flutter:
@@ -18,7 +18,7 @@ dev_dependencies:
   flutter_driver:
     sdk: flutter
   test: any
-  pedantic: ^1.9.2
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
