@@ -11,12 +11,11 @@ dev_dependencies:
   flutter_driver:
     sdk: flutter
   integration_test: ^1.0.2
-  pedantic: ^1.9.2
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  sdk: '>=2.12.0-259.9.beta <3.0.0'
   flutter: ">=1.9.1+hotfix.2"
-
