@@ -107,17 +107,17 @@ class DemoAppState extends State<DemoApp> {
     // RenderObject in its descendent tree when it's not
     // a RenderObjectWidget. The ElevatedButton's RenderObject
     // has its position and size after it's built.
-    final RenderBox box = context.findRenderObject();
+    final box = context.findRenderObject() as RenderBox?;
 
     if (imagePaths.isNotEmpty) {
       await Share.shareFiles(imagePaths,
           text: text,
           subject: subject,
-          sharePositionOrigin: box.localToGlobal(Offset.zero) & box.size);
+          sharePositionOrigin: box!.localToGlobal(Offset.zero) & box.size);
     } else {
       await Share.share(text,
           subject: subject,
-          sharePositionOrigin: box.localToGlobal(Offset.zero) & box.size);
+          sharePositionOrigin: box!.localToGlobal(Offset.zero) & box.size);
     }
   }
 }
